Update
Included functions to import IgBLAST annotations and export VDJ dataframe as AIRR-formatted TSV file in order to create the necessary input file for the IgPhyML tool. Included a parameter in the AntibodyForests function to import the IgPhyML output into an object of class 'AntibodyForests'. Included a script to synchronize the node names of two AntibodyForests objects.